Add scripts for two-stage reconcile of private->public->private push.

+++ tubular/git_repo.py
@@ -11,6 +11,7 @@ import re
 from six.moves import urllib
 from git import Repo
 from git.util import rmtree
+from git.exc import GitCommandError
 
 
 LOGGER = logging.getLogger(__name__)
@@ -20,6 +21,13 @@ LOGGER.setLevel(logging.INFO)
 class InvalidGitRepoURL(Exception):
     """
     Raised when repo URL can't be parsed.
+    """
+    pass
+
+
+class FastForwardMergeImpossible(Exception):
+    """
+    Raised when attempting a fast-forward-only merge that can't be done.
     """
     pass
 
@@ -70,11 +78,15 @@ class LocalGitAPI(object):
         )
         return cls(repo)
 
-    def push_branch(self, branch, remote='origin', force=False):
+    def push_branch(self, branch, remote='origin', remote_branch=None, force=False):
         """
-        Push a branch up to the remote server.
+        Push a branch up to the remote server, optionally with a different name.
         """
-        self.repo.remotes[remote].push('refs/heads/{}'.format(branch), force=force)
+        if remote_branch:
+            push_ref = 'refs/heads/{}:refs/heads/{}'.format(branch, remote_branch)
+        else:
+            push_ref = 'refs/heads/{}'.format(branch)
+        self.repo.remotes[remote].push(push_ref, force=force)
 
     def checkout_branch(self, branch):
         """
@@ -96,7 +108,12 @@ class LocalGitAPI(object):
             Commit SHA of the merge commit where the branch was merged.
         """
         self.checkout_branch(target_branch)
-        self.repo.git.merge(source_branch, ff_only=ff_only)
+        try:
+            self.repo.git.merge(source_branch, ff_only=ff_only)
+        except GitCommandError as exc:
+            if '--ff-only' in exc.command and exc.status == 128:
+                raise FastForwardMergeImpossible(' '.join(exc.command))
+            raise
         merge_sha = self.repo.git.rev_parse('HEAD')
         return merge_sha
 
@@ -108,6 +125,30 @@ class LocalGitAPI(object):
         remote = self.repo.create_remote(remote_name, remote_url)
         remote.fetch()
 
+    def get_head_sha(self, branch=None):
+        """
+        Gets the HEAD commit sha for the repo.
+        """
+        if branch:
+            return self.repo.heads[branch].commit.hexsha
+        else:
+            return self.repo.head.commit.hexsha
+
+    def create_branch(self, branch_name, commit='HEAD'):
+        """
+        Creates a branch with a specified name in the local repo.
+        """
+        self.repo.create_head(branch_name, commit)
+
+    def track_remote_branch(self, remote_name, remote_branch):
+        """
+        Create a local branch which tracks a remote branch.
+        """
+        branch_ref = self.repo.remote(remote_name).refs[remote_branch]
+        self.create_branch(remote_branch, branch_ref)
+        self.repo.heads[remote_branch].set_tracking_branch(branch_ref)
+        self.checkout_branch(remote_branch)
+
     def octopus_merge(self, base_branch, commitishes):
         """
         Merge all ``commitishes`` into ``base_branch`` in this repo.
@@ -115,7 +156,7 @@ class LocalGitAPI(object):
         self.checkout_branch(base_branch)
         if commitishes:
             self.repo.git.merge(*commitishes)
-        return self.repo.head.commit.hexsha
+        return self.get_head_sha()
 
     def force_branch_to(self, branch, commitish, remote=None):
         """
@@ -135,11 +176,12 @@ class LocalGitAPI(object):
             self.repo.heads[branch].reset(commitish, index=True)
 
     @contextmanager
-    def cleanup(self):
+    def cleanup(self, cleanup=True):
         """
         Delete the repo working directory when this contextmanager is finished.
         """
         try:
             yield self
         finally:
-            rmtree(self.repo.working_dir)
+            if cleanup:
+                rmtree(self.repo.working_dir)
